refactor: Observer 훅 분리
