+++ src/app/search/_components/Observer/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useObserver } from './useObserver';
+export { default as useFetchBooksAction } from './useFetchBooksAction';
